+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLmesh.h
@@ -15,3 +15,36 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
+
+#include <string>
+
+namespace enigma_user
+{
+
+unsigned int d3d_model_create();
+void d3d_model_destroy(const unsigned int id);
+bool d3d_model_exists(const unsigned int id);
+void d3d_model_clear(const unsigned int id);
+void d3d_model_save(const unsigned int id, std::string fname);
+bool d3d_model_load(const unsigned int id, std::string fname);
+void d3d_model_draw(const unsigned int id, double x, double y, double z, int texId);
+void d3d_model_primitive_begin(const unsigned int id, int kind);
+void d3d_model_primitive_end(const unsigned int id);
+void d3d_model_vertex(const unsigned int id, double x, double y, double z);
+void d3d_model_vertex_color(const unsigned int id, double x, double y, double z, int col, double alpha);
+void d3d_model_vertex_texture(const unsigned int id, double x, double y, double z, double tx, double ty);
+void d3d_model_vertex_texture_color(const unsigned int id, double x, double y, double z, double tx, double ty, int col, double alpha);
+void d3d_model_vertex_normal(const unsigned int id, double x, double y, double z, double nx, double ny, double nz);
+void d3d_model_vertex_normal_color(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, int col, double alpha);
+void d3d_model_vertex_normal_texture(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, double tx, double ty);
+void d3d_model_vertex_normal_texture_color(const unsigned int id, double x, double y, double z, double nx, double ny, double nz, double tx, double ty, int col, double alpha);
+void d3d_model_wall(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep);
+void d3d_model_floor(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep);
+void d3d_model_block(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep, bool closed = true);
+void d3d_model_cylinder(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep, bool closed, int steps);
+void d3d_model_cone(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep, bool closed, int steps);
+void d3d_model_ellipsoid(const unsigned int id, double x1, double y1, double z1, double x2, double y2, double z2, double hrep, double vrep, int steps);
+void d3d_model_icosahedron(const unsigned int id);
+void d3d_model_torus(const unsigned int id, double x1, double y1, double z1, double hrep, double vrep, int csteps, int tsteps, double radius, double tradius, double TWOPI = 2*3.14);
+
+}
