+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/include.h
@@ -4,6 +4,7 @@
 #include "../General/GLbackground.h"
 #include "../General/GLtextures.h"
 #include "../General/GLtiles.h"
+#include "GLmesh.h"
 
 #include "../General/GLfont.h"
 #include "../General/GLcurves.h"
